+++ R/functions_qtl.R
@@ -212,8 +212,7 @@ only_plot_profile <- function(pl.in){
 #' Pos_upper - upper position of the confidence interval; Pval - QTL p-value; h2 - herdability
 #' @param effects data.frame with: pheno - phenotype ID; qtl.id - QTL ID; haplo - haplotype ID; effect - haplotype effect value
 #' @param pheno.col integer identifying phenotype  
-#' @param p1 character with parent 1 ID
-#' @param p2 character with parent 2 ID
+#' @param parents vector with parents ID
 #' @param lgs vector of integers with linkage group ID of selected QTL/s
 #' @param groups vector of integers with selected linkage group ID
 #' @param position vector of centimorgan positions of selected QTL/s
@@ -231,9 +230,10 @@ only_plot_profile <- function(pl.in){
 #' 
 #' @keywords internal
 data_effects <- function(qtl_info, effects, pheno.col = NULL, 
-                         p1 = "P1", p2 = "P2",  
+                         parents = NULL,
                          lgs = NULL, groups = NULL, position = NULL, 
                          software, design = c("bar", "circle", "digenic")) {
+  
   
   CI.lower <- CI.upper <- x <- y <- z <- Estimates <- LG <- unique.id <- NULL
   x.axis <- haplo <- effect <- qtl.id <- Alleles <- . <- NULL
@@ -244,15 +244,43 @@ data_effects <- function(qtl_info, effects, pheno.col = NULL,
     pheno.col.n <- which(unique(qtl_info$pheno) %in% pheno.col)
   }
   
+  
   if(software == "QTLpoly" | software == "diaQTL"){
     
     if(software == "QTLpoly"){
       ploidy <- max(nchar(effects$haplo))
+      if(is.null(parents)) {# Multi-population still not implemented
+        p1 <- "P1"
+        p2 <- "P2" 
+      } else {
+        p1 <- parents[1]
+        p2 <- parents[2]
+      }
     } else if(software == "diaQTL") {
       get.size <-  filter(effects, .data$pheno == unique(qtl_info$pheno)[1] & .data$qtl.id == 1 & !grepl("x",.data$haplo)) # issue if parents name has x: fixme!
-      ploidy = nrow(get.size)/2
+      ploidy = as.numeric(table(substring(unique(get.size$haplo), 1, nchar(unique(get.size$haplo)) -2))[1])
+      
+      old.parents.names <- unique(substr(get.size$haplo,1,nchar(get.size$haplo)-2))
+      n.parents <- length(old.parents.names)
+      if(is.null(parents))  {
+        parents <- paste0("P", 1:n.parents)
+      } else {
+        if(length(parents) != n.parents)
+          stop(safeError(paste0("Your data set has", n.parents, " parental genotyopes. Please, provide a name for each one.")))
+      }
+      
+      # Update parents names in effects data
+      for(z in 1:n.parents)
+        effects$haplo <- gsub(old.parents.names[z], parents[z], effects$haplo)
     } else if(software == "polyqtlR"){
       ploidy <- (dim(effects)[2] - 3)/2
+      if(is.null(parents)) {# Multi-population still not implemented
+        p1 <- "P1"
+        p2 <- "P2" 
+      } else {
+        p1 <- parents[1]
+        p2 <- parents[2]
+      }    
     }
     
     qtl_info.sub <- qtl_info %>% filter(.data$pheno %in% unique(qtl_info$pheno)[pheno.col.n]) %>%
@@ -287,9 +315,9 @@ data_effects <- function(qtl_info, effects, pheno.col = NULL,
           if(ploidy == 4) {
             if(software == "diaQTL"){
               if(any(data$type == "Digenic")){
-                data <- data.frame(Estimates=as.numeric(data$effect), CI.lower = data$CI.lower, CI.upper = data$CI.upper, Alleles=data$haplo, Parent=c(rep(p1,4),rep(p2,4),rep(p1,14),rep(p2,14)), Effects=c(rep("Additive",8),rep("Digenic",28)))
+                data <- data.frame(Estimates=as.numeric(data$effect), CI.lower = data$CI.lower, CI.upper = data$CI.upper, Alleles=data$haplo, Parent=c(rep(parents, each = ploidy),rep(NA,dim(data)[1]-n.parents*ploidy)), Effects=c(rep("Additive",n.parents*ploidy),rep("Digenic",dim(data)[1]-n.parents*ploidy)))
               } else  {
-                data <- data.frame(Estimates=as.numeric(data$effect), CI.lower = data$CI.lower, CI.upper = data$CI.upper, Alleles=data$haplo, Parent=c(rep(p1,4),rep(p2,4)), Effects=c(rep("Additive",8)))
+                data <- data.frame(Estimates=as.numeric(data$effect), CI.lower = data$CI.lower, CI.upper = data$CI.upper, Alleles=data$haplo, Parent=rep(parents, each = ploidy), Effects="Additive")
               }
             } else {
               data <- data[1:36,]
@@ -326,7 +354,9 @@ data_effects <- function(qtl_info, effects, pheno.col = NULL,
           if(design == "bar"){
             if(software == "QTLpoly"){
               lim <- max(abs(data[which(data$Effects == "Additive"),]$Estimates))
-            } else lim <- max(abs(c(data[which(data$Effects == "Additive"),]$CI.lower, data[which(data$Effects == "Additive"),]$CI.upper)))
+            } else 
+              lim <- max(abs(c(data[which(data$Effects == "Additive"),]$CI.lower, data[which(data$Effects == "Additive"),]$CI.upper)))
+            
             plot <- ggplot(data[which(data$Effects == "Additive"),], aes(x = Alleles, y = Estimates, fill = Estimates)) +
               geom_bar(stat="identity") + ylim(c(-lim, lim)) +
               {if(software == "diaQTL") geom_errorbar(aes(ymin=CI.lower, ymax=CI.upper), width=.2, position=position_dodge(.9))} +
@@ -339,6 +369,7 @@ data_effects <- function(qtl_info, effects, pheno.col = NULL,
                     plot.subtitle = element_text(hjust = 0.5), 
                     axis.text.x.bottom = element_text(hjust = 1, vjust = 0.5))
             plots1[[q]] <- plot
+            
           } else if(design == "digenic"){
             if(!all(is.na(data[which(data$Effects == "Digenic"),]$Estimates))){
               temp <- do.call(rbind, strsplit(data$Alleles, "x"))
@@ -358,6 +389,7 @@ data_effects <- function(qtl_info, effects, pheno.col = NULL,
                                                additive.effects$Estimates[match(digenic.effects$x, additive.effects$Alleles)] + 
                                                additive.effects$Estimates[match(digenic.effects$y, additive.effects$Alleles)]))
               }
+              
               plot = ggplot(data= plot.data,aes(x= x, y= y, fill= z)) + 
                 geom_tile() + scale_fill_gradient2(name="") + 
                 labs(title = paste("Trait:", unique(qtl_info$pheno)[p]),
@@ -368,6 +400,7 @@ data_effects <- function(qtl_info, effects, pheno.col = NULL,
                 coord_fixed(ratio=1)
               plots1[[q]] <- plot
             } else plots1[[q]] <- NULL
+            
           } else if(design == "circle"){
             additive.effects <- data[which(data$Effects == "Additive"),]
             additive.effects$pheno <- unique(qtl_info$pheno)[p]
@@ -727,6 +760,7 @@ plot.mappoly.homoprob <- function(x, stack = FALSE, lg = NULL,
 #' Plot selected haplotypes
 #' 
 #' @param input.haplo character vector with selected haplotypes. It contains the information: "Trait:<trait ID>_LG:<linkage group ID_Pos:<QTL position>" 
+#' @param exclude.haplo character vector with haplotypes to be excluded. It contains the information: "Trait:<trait ID>_LG:<linkage group ID_Pos:<QTL position>" 
 #' @param probs data.frame with first column (named `ind`) as individuals ID and next columns named with markers ID and containing the genotype probability at each marker
 #' @param selected_mks data.frame with: LG - linkage group ID; mk - marker ID; pos - position in linkage map (cM)
 #' @param effects.data output object from \code{data_effects} function
@@ -735,9 +769,10 @@ plot.mappoly.homoprob <- function(x, stack = FALSE, lg = NULL,
 #' 
 #' 
 #' @keywords internal
-select_haplo <- function(input.haplo, probs, selected_mks, effects.data){
+select_haplo <- function(input.haplo,probs, selected_mks, effects.data, exclude.haplo = NULL){
   LG <- map.position <- individual <- probability <- NULL
   
+  # Include haplo
   lgs <- sapply(strsplit(unlist(input.haplo), "_"),function(x) x[grep("LG", x)])
   lgs <- gsub("LG:", "", lgs)
   homo.dat <- calc_homologprob(probs = probs, selected_mks = selected_mks, selected_lgs = lgs)
@@ -761,18 +796,50 @@ select_haplo <- function(input.haplo, probs, selected_mks, effects.data){
     like.ind.all[[i]] <-  like.ind
   }
   like.intersect <- Reduce(intersect, like.ind.all)
+  
+  ## Exclude haplo
+  if(!is.null(exclude.haplo)){
+    lgs1 <- sapply(strsplit(unlist(exclude.haplo), "_"),function(x) x[grep("LG", x)])
+    lgs1 <- gsub("LG:", "", lgs1)
+    homo.dat1 <- calc_homologprob(probs = probs, selected_mks = selected_mks, selected_lgs = lgs1)
+    pos1 <- sapply(strsplit(unlist(exclude.haplo), "_"),function(x) x[grep("Pos", x)])
+    pos1 <- gsub("Pos:", "", pos1)
+    homo <- sapply(strsplit(unlist(exclude.haplo), "_"),function(x) x[grep("homolog", x)])
+    homo <- gsub("homolog:", "", homo)
+    alleles <- effects.data[[1]]$data$Alleles[!grepl("_",effects.data[[1]]$data$Alleles)]
+    alleles <- rep(alleles, length(homo))
+    like.ind.all <- list()
+    for(i in 1:length(pos1)){
+      idx <- match(homo[i], sort(unique(alleles)))
+      homoprob_temp <- homo.dat1$homoprob %>% 
+        filter(round(map.position,2) %in% round(as.numeric(pos1[i]),2)) %>% filter(LG %in% lgs1[i])
+      homoprob_temp <- homoprob_temp[order(homoprob_temp$individual, homoprob_temp$homolog),]
+      homoprob_temp <- homoprob_temp %>% 
+        group_by(map.position, LG, individual) %>% 
+        summarise(best = which(probability > 0.5))
+      like.ind <-  homoprob_temp$individual[which(homoprob_temp$best %in% idx)]
+      if(length(like.ind) ==0) like.ind <- NA
+      like.ind.all[[i]] <-  like.ind
+    }
+    like.intersect.exclude <- Reduce(intersect, like.ind.all)
+    like.intersect <- like.intersect[-which(like.intersect %in% like.intersect.exclude)]
+    # For vertical lines
+    idx <- which(paste0(round(homo.dat$homoprob$map.position,2), "_", homo.dat$homoprob$LG) %in% c(paste0(round(as.numeric(pos),2), "_", lgs), paste0(round(as.numeric(pos1),2), "_", lgs1)))
+  } else {
+    idx <- which(paste0(round(homo.dat$homoprob$map.position,2), "_", homo.dat$homoprob$LG) %in% paste0(round(as.numeric(pos),2), "_", lgs))
+  }
+  
   if(length(like.intersect) == 0 | all(is.na(like.intersect))) stop(safeError("No individual in the progeny was found containing the combination of all the selected homolog/s. Please, select another combination of homolog/s."))
-  idx <- which(paste0(round(homo.dat$homoprob$map.position,2), "_", homo.dat$homoprob$LG) %in% paste0(round(as.numeric(pos),2), "_", lgs))
   homo.dat$homoprob$qtl <- NA
   homo.dat$homoprob$qtl[idx] <- homo.dat$homoprob$map.position[idx] # vertical lines
   
   p <- list()
   for(i in 1:length(like.intersect)){
     p[[i]] <- plot.mappoly.homoprob(x = homo.dat, 
-                                    lg = unique(as.numeric(lgs)), 
-                                    ind = as.character(like.intersect)[i],
-                                    use.plotly = FALSE)
+                                               lg = unique(as.numeric(lgs)), 
+                                               ind = as.character(like.intersect)[i],
+                                               use.plotly = FALSE)
   }
-  return(p)
+  return(list(p, inds = as.character(like.intersect)))
 }