+++ R/mod_qtl_view.R
@@ -78,7 +78,7 @@ mod_qtl_view_ui <- function(id){
                                    downloadButton(ns('bn_download'), "Download", class = "butt")
                             ),
                             column(3,
-                                   radioButtons(ns("fformat"), "File type", choices=c("png","tiff","jpeg","pdf"), selected = "png", inline = T)
+                                   radioButtons(ns("fformat"), "File type", choices=c("png","tiff","jpeg","pdf", "RData"), selected = "png", inline = T)
                             ),                     
                             column(2,
                                    numericInput(ns("width_profile"), "Width (mm)", value = 180),
@@ -110,18 +110,24 @@ mod_qtl_view_ui <- function(id){
                          column(3,
                                 downloadButton(ns('bn_download_effects'), "Download", class = "butt")
                          ),
+                         
                          column(3,
-                                radioButtons(ns("fformat_effects"), "File type", choices=c("png","tiff","jpeg","pdf"), selected = "png", inline = T)
-                         ),                     
-                         column(2,
                                 numericInput(ns("width_effects"), "Width (mm)", value = 180),
                          ),
-                         column(2,
+                         column(3,
                                 numericInput(ns("height_effects"), "Height (mm)", value = 120),
                          ),
-                         column(2,
+                         column(3,
                                 numericInput(ns("dpi_effects"), "DPI", value = 300)
                          ), br(), 
+                         column(12, 
+                                column(6,
+                                       radioButtons(ns("fformat_effects"), "File type", choices=c("png","tiff","jpeg","pdf", "RData"), selected = "png", inline = T)
+                                ),
+                                column(6,
+                                       textInput(ns("parents_name"), "Parents name", value = "P1, P2")
+                                ),
+                         ),
                          column(12,
                                 hr(),
                                 uiOutput(ns("plot.ui"))
@@ -151,13 +157,26 @@ mod_qtl_view_ui <- function(id){
                                               dropdownAlignRight = TRUE
                                             ), 
                                             multiple = TRUE), br(),
+                                pickerInput(ns("haplo_exclude"),
+                                            label = h6("Exclude haplotypes (optional)"),
+                                            choices = "Click on `update available haplotype` to update",
+                                            selected = "Click on `update available haplotype` to update",
+                                            options = pickerOptions(
+                                              size = 15,
+                                              `selected-text-format` = "count > 3",
+                                              `live-search`=TRUE,
+                                              actionsBox = TRUE,
+                                              dropupAuto = FALSE,
+                                              dropdownAlignRight = TRUE
+                                            ), 
+                                            multiple = TRUE), br(),
                                 actionBttn(ns("haplo_submit"), style = "jelly", color = "royal",  size = "sm", label = "submit selected haplotypes*", icon = icon("share-square", verify_fa = FALSE)), 
                                 br(), hr()),
                          column(3,
                                 downloadButton(ns('bn_download_haplo'), "Download", class = "butt")
                          ),
                          column(3,
-                                radioButtons(ns("fformat_haplo"), "File type", choices=c("png","tiff","jpeg","pdf"), selected = "png", inline = T)
+                                radioButtons(ns("fformat_haplo"), "File type", choices=c("png","tiff","jpeg","pdf", "RData"), selected = "png", inline = T)
                          ),                     
                          column(2,
                                 numericInput(ns("width_haplo"), "Width (mm)", value = 180),
@@ -170,6 +189,7 @@ mod_qtl_view_ui <- function(id){
                          ), br(), 
                          column(12,
                                 hr(),
+                                htmlOutput(ns("ind_names")), hr(),
                                 uiOutput(ns("plot_haplo.ui"))
                          )
                      ),
@@ -288,8 +308,8 @@ mod_qtl_view_server <- function(input, output, session,
   
   qtl.data <- reactive({
     validate(
-      need(length(input$phenotypes) != 0, "Select at least one phenotype"),
-      need(length(input$group) != 0, "Select at least one linkage group"),
+      need(length(input$phenotypes) != 0 & input$phenotypes != "Upload QTL information to update", "Select at least one phenotype"),
+      need(length(input$group) != 0 & input$group != "Upload map or QTL data in `upload` session.", "Select at least one linkage group"),
       need(!is.null(loadQTL()), "Upload the QTL information in upload session to access this feature.")
     )
     idx <- which(unique(loadQTL()$profile$pheno) %in% input$phenotypes)
@@ -321,6 +341,9 @@ mod_qtl_view_server <- function(input, output, session,
     validate(
       need(dim(df)[1] > 0, "Select at least one triangle on the bottom of the QTL profile graphic. The triangles refer to QTL peaks detected. You can click and brush your cursor to select more than one.")
     )
+    
+    parents <- unlist(strsplit(input$parents_name, ","))
+    
     withProgress(message = 'Working:', value = 0, {
       incProgress(0.5, detail = paste("Getting data..."))
       data <- data_effects(qtl_info = loadQTL()$qtl_info, 
@@ -330,7 +353,8 @@ mod_qtl_view_server <- function(input, output, session,
                            position = df$`Position (cM)`,
                            groups = as.numeric(input$group),
                            software = loadQTL()$software,
-                           design = input$effects_design)
+                           design = input$effects_design,
+                           parents = parents)
     })
   })
   
@@ -340,7 +364,6 @@ mod_qtl_view_server <- function(input, output, session,
       plot_effects(effects.data(), software = loadQTL()$software, design = input$effects_design)
     })
   })
-  
   
   plotHeight <- reactive({
     
@@ -383,12 +406,22 @@ mod_qtl_view_server <- function(input, output, session,
                           label = "Select haplotypes",
                           choices = paste0("Feature not implemented for software: ", loadQTL()$software),
                           selected= paste0("Feature not implemented for software: ", loadQTL()$software))
+        
+        updatePickerInput(session, "haplo_exclude",
+                          label = "Exclude haplotypes (optional)",
+                          choices = paste0("Feature not implemented for software: ", loadQTL()$software),
+                          selected= paste0("Feature not implemented for software: ", loadQTL()$software))
       } else if(!is.null(input$plot_brush)){
         dframe <- brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat")
       } else {
         dframe <- NULL
         updatePickerInput(session, "haplo",
                           label = "Select haplotypes",
+                          choices = "Select QTL in the profile graphic to update",
+                          selected= "Select QTL in the profile graphic to update")
+        
+        updatePickerInput(session, "haplo_exclude",
+                          label = "Exclude haplotypes (optional)",
                           choices = "Select QTL in the profile graphic to update",
                           selected= "Select QTL in the profile graphic to update")
       }
@@ -398,12 +431,21 @@ mod_qtl_view_server <- function(input, output, session,
                         label = "Select haplotypes",
                         choices = "Upload the QTL information in upload session to access this feature.",
                         selected= "Upload the QTL information in upload session to access this feature.")
+      
+      updatePickerInput(session, "haplo_exclude",
+                        label = "Exclude haplotypes (optional)",
+                        choices = "Upload the QTL information in upload session to access this feature.",
+                        selected= "Upload the QTL information in upload session to access this feature.")
     }
-    
     if(!is.null(dframe)){
       if(input$effects_design == "digenic" | input$effects_design == "circle") {
         updatePickerInput(session, "haplo",
                           label = "Select haplotypes",
+                          choices = "Select `bar` design to access this feature.",
+                          selected= "Select `bar` design to access this feature.")
+        
+        updatePickerInput(session, "haplo_exclude",
+                          label = "Exclude haplotypes (optional)",
                           choices = "Select `bar` design to access this feature.",
                           selected= "Select `bar` design to access this feature.")
       } else {
@@ -418,6 +460,11 @@ mod_qtl_view_server <- function(input, output, session,
                           label = "Select haplotypes",
                           choices = haplo_choices,
                           selected= haplo_choices[1:3])
+        
+        updatePickerInput(session, "haplo_exclude",
+                          label = "Exclude haplotypes (optional)",
+                          choices = haplo_choices,
+                          selected= NULL)
       }
     }
   })
@@ -429,11 +476,14 @@ mod_qtl_view_server <- function(input, output, session,
       need(all(input$haplo != "Select QTL in the profile graphic to update"), "Select QTL in the profile graphic to update"),
       need(all(input$haplo != "Select `bar` design to access this feature."), "Select `bar` design to access this feature.")
     )
-    p <- select_haplo(input$haplo, loadQTL()$probs, loadQTL()$selected_mks, effects.data())
+    
+    list.p <- select_haplo(input$haplo, loadQTL()$probs, loadQTL()$selected_mks, effects.data(), exclude.haplo = input$haplo_exclude)
+    p <- list.p[[1]]
+    inds <- list.p[[2]]
     counts <- ceiling(length(p)/3)
     if(counts == 0) counts <- 1
     size <- counts*450
-    list(p, size)
+    list(p, size, inds)
   })
   
   output$haplotypes <- renderPlot({
@@ -448,6 +498,11 @@ mod_qtl_view_server <- function(input, output, session,
   
   output$plot_haplo.ui <- renderUI({
     plotOutput(ns("haplotypes"), height = haplo_data()[[2]])
+  })
+  
+  output$ind_names <- renderUI({
+    x <- paste0("<strong>Number of individuals with selected haplotypes: ",length(haplo_data()[[3]]),"   ","<br><strong>Individual's ID  </strong>: ", paste(haplo_data()[[3]], collapse = ", "))
+    HTML(x)
   })
   
   output$info <- DT::renderDataTable(server = FALSE, {
@@ -508,6 +563,7 @@ mod_qtl_view_server <- function(input, output, session,
     if(input$fformat=="tiff") filename <- paste0("profile","_",seed,".tiff")
     if(input$fformat=="jpeg") filename <- paste0("profile","_",seed,".jpg")
     if(input$fformat=="pdf") filename <- paste0("profile","_",seed,".pdf")
+    if(input$fformat=="RData") filename <- paste0("profile","_",seed,".RData")
     return(filename)
   })
   
@@ -515,8 +571,11 @@ mod_qtl_view_server <- function(input, output, session,
   fn_download <- function()
   {
     p <- only_plot_profile(pl.in = qtl.data())
-    ggsave(p, filename = fn_downloadname(), 
-           width = input$width_profile, height = input$height_profile, units = "mm", dpi = input$dpi_profile)    
+    
+    if(input$fformat!="RData"){
+      ggsave(p, filename = fn_downloadname(), 
+             width = input$width_profile, height = input$height_profile, units = "mm", dpi = input$dpi_profile)    
+    } else save(p, file = fn_downloadname())
   }
   
   observe({
@@ -548,6 +607,7 @@ mod_qtl_view_server <- function(input, output, session,
     if(input$fformat_effects=="tiff") filename <- paste0("effects","_",seed,".tiff")
     if(input$fformat_effects=="jpeg") filename <- paste0("effects","_",seed,".jpg")
     if(input$fformat_effects=="pdf") filename <- paste0("effects","_",seed,".pdf")
+    if(input$fformat_effects=="RData") filename <- paste0("effects","_",seed,".RData")
     return(filename)
   })
   
@@ -560,10 +620,16 @@ mod_qtl_view_server <- function(input, output, session,
     
     df <- brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat")
     
+    if(!grepl("Example" ,input$parents_name)) {
+      cat("here2")
+      parents <- unlist(strsplit(input$parents_name, ","))
+    } else parents <- NULL
+    
     data <- data_effects(qtl_info = loadQTL()$qtl_info, 
                          effects = loadQTL()$effects,
                          pheno.col = as.character(df$Trait), 
                          lgs = df$LG, 
+                         parents = parents,
                          position = df$`Position (cM)`,
                          groups = as.numeric(input$group),
                          software = loadQTL()$software,
@@ -571,8 +637,10 @@ mod_qtl_view_server <- function(input, output, session,
     
     plots <- plot_effects(data, software = loadQTL()$software, design = input$effects_design)
     
-    ggsave(plots, filename = fn_downloadname_effects(), height = input$height_effects, 
-           width = input$width_effects, units = "mm", bg = "white", dpi = input$dpi_effects)    
+    if(input$fformat_effects!="RData"){
+      ggsave(plots, filename = fn_downloadname_effects(), height = input$height_effects, 
+             width = input$width_effects, units = "mm", bg = "white", dpi = input$dpi_effects)    
+    } else save(data, file = fn_downloadname_effects())
   }
   
   shinyjs::disable("bn_download_effects")
@@ -614,6 +682,7 @@ mod_qtl_view_server <- function(input, output, session,
     if(input$fformat_haplo=="tiff") filename <- paste0("haplotypes","_",seed,".tiff")
     if(input$fformat_haplo=="jpeg") filename <- paste0("haplotypes","_",seed,".jpg")
     if(input$fformat_haplo=="pdf") filename <- paste0("haplotypes","_",seed,".pdf")
+    if(input$fformat_haplo=="RData") filename <- paste0("haplotypes","_",seed,".RData")
     return(filename)
   })
   
@@ -623,8 +692,10 @@ mod_qtl_view_server <- function(input, output, session,
     p <- select_haplo(input$haplo, loadQTL()$probs, loadQTL()$selected_mks, effects.data())
     plots <- ggarrange(plotlist = p, ncol = 3, common.legend = TRUE)
     
-    ggsave(plots, filename = fn_downloadname_haplo(), height = input$height_haplo, 
-           width = input$width_haplo, units = "mm", bg = "white", dpi = input$dpi_haplo)    
+    if(input$fformat_haplo!="RData"){
+      ggsave(plots, filename = fn_downloadname_haplo(), height = input$height_haplo, 
+             width = input$width_haplo, units = "mm", bg = "white", dpi = input$dpi_haplo)    
+    } else save(p, file = fn_downloadname_haplo())
   }
   
   observe({